chore: Refactored the account settings input layout
