+++ src/pages/dashboard/actions/actions.js
@@ -1,0 +1,11 @@
+export const actions = {
+  inputChanged: 'input_change',
+  formSubmitted: 'submit_form',
+  isSubmitting: 'is_submitting',
+  isNotSubmitting: 'is_not_submitting',
+  isFormEmpty: 'is_form_empty',
+  toggleSuccessMessage: 'toggled_success_message',
+  isPasswordMatch: 'password_matched',
+  submitted: 'submitted',
+  cancelForm: 'form_canceled',
+};
