+++ src/pages/dashboard/styles/dashboard.css
@@ -2,8 +2,7 @@
 
 @layer components {
   .container-dashboard {
-    @apply mx-auto lg:mx-0 px-4 md:px-0;
-    width: clamp(375px, 90vw, 650px);
+    @apply mx-auto lg:mx-0 w-[90%] lg:max-w-5xl;
   }
 
   .dashboardSidebarLinks {
